# Optimize IOrderedEnumerable.Take(1)