@AndreyAkinshin 2018-09-23
@danmosemsft, it seems that we don't have such tests. Should I add them in `TakeTests.cs`/`TakeLastTests.cs` or `OrderByTests.cs`?