@danmosemsft 2018-09-23
Your benchmark does not seem to cover Take Last. I guess that is self evidently the same though