@danmosemsft 2018-09-23
You might also consider whether it is worth adding your performance benchmark into the repo for regression protection.