@AndreyAkinshin 2018-09-24
@danmosemsft, I updated the PR. Now it contains the correctness tests and performance tests. It turned out that `array.OrderBy(x => x).TakeLast(1)` doesn't use `QuickSelect`, so I removed the corresponding logic. In this PR, only `Take(1)` is optimized.