@stephentoub 2018-09-24
Unless I'm missing it, I only see tests here for Take(1), which is the case this optimization is meant to improve.  What about other N values, where this optimization won't help?  Does it hurt at all?

I'm not against the optimization if there's no measurable impact on other cases; I just question how valuable this actually is... after all, a developer doing OrderBy(...).Take(1) could more easily write Min(...) themselves.