@AndreyAkinshin 2018-09-24
@stephentoub, Originally, I tried to optimize the `Take` method in general with the help of the partial heapsort. It has an awesome performance boost in the worst case, but it slows down the average case, so we need a better solution.
During benchmarking, I found out that `Take(1)` uses `QuickSelect` instead of `QuickSort` (we handle the case of a single element [here](https://github.com/dotnet/corefx/blob/v2.2.0-preview2/src/System.Linq/src/System/Linq/OrderedEnumerable.cs#L90)) I don't know why this case is so special: it seems that we can use `QuickSort` for `ElementAt` as well, it works the same way with the same time complexity.
Anyway, I already had a proven optimization for this small case (without any side-effects or impact on other cases), so I decided to create a pull request. In real life, such construction can be used as a special case of `Take(K)`.