@AndreyAkinshin 2018-09-26
@stephentoub 
> It's unlikely to affect anything measurably, but in the name of completeness, you've validated that this extra check / the extra code in the method doesn't negatively impact cases beyond the positive cases you created benchmarks for?

I improved the Welch's t-test in BenchmarkDotNet, now it's possible to specify a threshold for distribution comparing (the update is available in v0.11.1.754-nightly; it will be included in v0.11.2). Here is my benchmark:

```cs
public class MyConfig : ManualConfig
{
    public MyConfig()
    {
        var baseJob = Job.Default
            .WithWarmupCount(10)
            .WithIterationCount(300)
            .WithInvocationCount(2097152)
            .WithOutlierMode(OutlierMode.None);
        Add(baseJob.With(new CoreRunToolchain(new FileInfo(
                "/home/akinshin/Work/dotnet/corefx-upstream/bin/runtime/netcoreapp-Linux-Debug-x64/corerun")))
            .WithId("c8e3ba6c09")
            .AsBaseline()
        );
        Add(baseJob.With(new CoreRunToolchain(new FileInfo(
                "/home/akinshin/Work/dotnet/corefx/bin/runtime/netcoreapp-Linux-Debug-x64/corerun")))
            .WithId("corefx#32389")
        );
    }
}

[Config(typeof(MyConfig))]
[WelchTTestAbsoluteNanoseconds(0.3)]
[RPlotExporter]
public class TakeOverheadBenchmark
{
    private int[] array = {1, 2, 3};

    [Benchmark]
    public int Take1()
    {
        int sum = 0;
        foreach (var x in array.OrderBy(x => x).Take(1))
            sum += x;
        return sum;
    }

    [Benchmark]
    public int Take2()
    {
        int sum = 0;
        foreach (var x in array.OrderBy(x => x).Take(2))
            sum += x;
        return sum;
    }
}
```
The `corefx-upstream` folder contains the `c8e3ba6c09` revision; the `corefx` folder contains `c8e3ba6c09` plus the patch from this PR. The iteration count is 300 to provide better statistics metrics for distributions. The invocation count is fixed (`2097152`) to provide the same accuracy between different runtimes. Here are my results:

```

BenchmarkDotNet=v0.11.1.754-nightly, OS=ubuntu 18.04
Intel Core i7-7700K CPU 4.20GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.1.402
  [Host]       : .NET Core 2.1.4 (CoreCLR 4.6.26814.03, CoreFX 4.6.26814.02), 64bit RyuJIT
  c8e3ba6c09   : .NET Core ? (CoreCLR 4.6.26920.03, CoreFX 4.6.27019.0), 64bit RyuJIT
  corefx#32389 : .NET Core ? (CoreCLR 4.6.26920.03, CoreFX 4.6.27019.0), 64bit RyuJIT

OutlierMode=None  Toolchain=CoreRun  InvocationCount=2097152  
IterationCount=300  WarmupCount=10  

 Method |          Job |     Mean |     Error |   StdDev |   Median | Scaled | ScaledSD | p-value |
------- |------------- |---------:|----------:|---------:|---------:|-------:|---------:|--------:|
  Take1 |   c8e3ba6c09 | 325.1 ns | 0.7870 ns | 4.101 ns | 323.7 ns |   1.00 |     0.00 |  1.0000 |
  Take1 | corefx#32389 | 315.5 ns | 0.4522 ns | 2.357 ns | 314.7 ns |   0.97 |     0.01 |  0.0000 |
        |              |          |           |          |          |        |          |         |
  Take2 |   c8e3ba6c09 | 350.3 ns | 0.7037 ns | 3.667 ns | 349.1 ns |   1.00 |     0.00 |  1.0000 |
  Take2 | corefx#32389 | 348.6 ns | 0.9457 ns | 4.929 ns | 347.2 ns |   1.00 |     0.02 |  1.0000 |

// * Legends *
  Mean     : Arithmetic mean of all measurements
  Error    : Half of 99.9% confidence interval
  StdDev   : Standard deviation of all measurements
  Median   : Value separating the higher half of all measurements (50th percentile)
  Scaled   : Mean(CurrentBenchmark) / Mean(BaselineBenchmark)
  ScaledSD : Standard deviation of ratio of distribution of [CurrentBenchmark] and [BaselineBenchmark]
  p-value  : p-value for Welch's t-test against baseline (H1: True difference in means > 0.3 ns)
  1 ns     : 1 Nanosecond (0.000000001 sec)
```

This benchmark was executed several times on 3 different machines (Linux/Windows/macOS with different hardware configurations). All the time I have the same conclusion based on the Welch's t-test: we can't detect statistically significant difference which is more than 0.3ns between mean values of `c8e3ba6c09` and `corefx#32389` for `Take2`. There are the density and timeline plots for this case:
![consoleapp11 takeoverheadbenchmark-take2-density](https://user-images.githubusercontent.com/2259237/46069346-0c3d7c00-c184-11e8-8987-0785e9795903.png)
![consoleapp11 takeoverheadbenchmark-take2-facettimelinesmooth](https://user-images.githubusercontent.com/2259237/46069407-2bd4a480-c184-11e8-8b45-a8f6c24ee8aa.png)